backend/app/api/users.py
"""
Users API Router
User management and profile endpoints
"""

from fastapi import APIRouter, Depends, HTTPException, status
from typing import List
import sys
import os

# Add schemas to Python path
sys.path.append(os.path.join(os.path.dirname(__file__), "..", "..", "..", "schemas"))

from pydantic_models import User, UserCreate, UserUpdate

router = APIRouter()


@router.get("/", response_model=List[User])
async def get_users():
    """Get all users (admin only)"""
    # TODO: Implement user listing with proper authentication
    return []


@router.get("/me", response_model=User)
async def get_current_user():
    """Get current user profile"""
    # TODO: Implement current user retrieval
    raise HTTPException(
        status_code=status.HTTP_501_NOT_IMPLEMENTED,
        detail="Endpoint not implemented yet"
    )


@router.get("/{user_id}", response_model=User)
async def get_user(user_id: str):
    """Get user by ID"""
    # TODO: Implement user retrieval by ID
    raise HTTPException(
        status_code=status.HTTP_501_NOT_IMPLEMENTED,
        detail="Endpoint not implemented yet"
    )


@router.post("/", response_model=User, status_code=status.HTTP_201_CREATED)
async def create_user(user: UserCreate):
    """Create new user"""
    # TODO: Implement user creation
    raise HTTPException(
        status_code=status.HTTP_501_NOT_IMPLEMENTED,
        detail="Endpoint not implemented yet"
    )


@router.put("/{user_id}", response_model=User)
async def update_user(user_id: str, user_update: UserUpdate):
    """Update user profile"""
    # TODO: Implement user update
    raise HTTPException(
        status_code=status.HTTP_501_NOT_IMPLEMENTED,
        detail="Endpoint not implemented yet"
    )


@router.delete("/{user_id}", status_code=status.HTTP_204_NO_CONTENT)
async def delete_user(user_id: str):
    """Delete user account"""
    # TODO: Implement user deletion
    raise HTTPException(
        status_code=status.HTTP_501_NOT_IMPLEMENTED,
        detail="Endpoint not implemented yet"
    )